requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
PyMySQL==0.6.3
SQLAlchemy==0.9.8
Tempita==0.5.2
Wand==0.3.8
Werkzeug==0.9.6
beautifulsoup4==4.3.2
decorator==3.4.0
itsdangerous==0.24
pbr==0.10.2
requests==2.5.0
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.14
